Simplify flask_app
