--- flask_app/templates/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <title>Search idol stars</title>
-</head>
-<body>
-<h1><a href="/upload">upload</a></h1>
-</body>
-</html>
