--- flask_app/templates/upload.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <title>Upload</title>
-</head>
-<body>
-<h1>Try to find your star</h1>
-<form action="/upload" enctype="multipart/form-data" method="POST">
-    <input name="file" type="file"/>
-    <input type="submit"/>
-</form>
-<div>
-    {{ matches }}
-    <hr>Message: {{ message }}
-</div>
-</body>
-</html>
